address comments: move startup triggering logic to ArcsShellApi
